Separate generate script to use generateCore for testing and add tests
